[FIX] graphql_base: rename types.py
This conflicts with the build system, as
the project root directory is in sys.path when building and
there is confusion with the stdlib types module.
