+++ graphql_base/__init__.py
@@ -2,4 +2,4 @@
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 
 from .controllers import GraphQLControllerMixin
-from .types import OdooObjectType
+from .graphene_types import OdooObjectType, odoo_attr_resolver
